extracted pop out test
