+++ src/Avalonia.Markup.Declarative.Tests/ControlsTests/ExternalPropertySetTests.cs
@@ -65,27 +65,6 @@ public class ExternalPropertySetTest(ITestOutputHelper testOutputHelper) : Avalo
     }
 
     [Fact]
-    public void ExternalPropertySetTestView_InnerValueChangePoppedToParentComponent()
-    {
-        var view = new ExternalPropertySetTestView();
-
-        var window = new Window { Content = view };
-        window.Show();
-        Dispatcher.UIThread.RunJobs();
-
-        Assert.Equal(1, view.Value);
-
-        var expectedValue = 50;
-        view.SliderWithLabel.ChangeSliderValue(expectedValue);
-
-        var sliderWithLabelValue = view.SliderWithLabel.Value;
-        Assert.Equal(expectedValue, sliderWithLabelValue);
-
-        var realValue = view.Value;
-        Assert.Equal(expectedValue, realValue);
-    }
-
-    [Fact]
     public void ExternalPropertySetTestView_BindPropertyChangeInnerValue()
     {
         var view = new ExternalPropertySetTestView();
